% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savepointr.R
\name{print.savepointr}
\alias{print.savepointr}
\title{Print a savepointr object.}
\usage{
\method{print}{savepointr}(x, ...)
}
\arguments{
\item{x}{A \code{\link{savepointr}}.}

\item{...}{Not used.}
}
\description{
Print a savepointr object.
}
