% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savepointr.R
\name{savepointr}
\alias{savepointr}
\alias{current_savepoint}
\alias{has_current_savepoint}
\alias{clear_savepointr}
\alias{save_current}
\title{Create a savepointr}
\usage{
savepointr(directory = ".", create_directory = TRUE,
  current_path = file.path(directory, "savepoint.rds"),
  new_path = file.path(directory, "new_savepoint.rds"))

current_savepoint(sptr)

has_current_savepoint(sptr)

clear_savepointr(sptr)

save_current(sptr, savepoint)
}
\arguments{
\item{directory}{The directory in which to store the savepoint. Ignored if
current_path and new_path are provided.}

\item{create_directory}{If TRUE, and directory is not ignored, will create
the directory if it does not yet exist.}

\item{current_path}{Path at which the current savepoint will be stored.
Regardless of the file extension, will be stored in RDS format.}

\item{new_path}{Path where new savepoints will initially be saved, before
being safely renamed over the current path.}

\item{sptr}{A savepointr object.}

\item{savepoint}{An arbitrary R object (must be serializable).}
}
\description{
Creates a savepointr in the specified directory.
}
\section{Functions}{
\itemize{
\item \code{current_savepoint}: Return the current savepoint, if any

\item \code{has_current_savepoint}: Whether a current savepoint exists, TRUE/FALSE

\item \code{clear_savepointr}: Clear the current savepoint, if any, and any failed
saves

\item \code{save_current}: Create a new savepoint
}}

